Move `AclForDb` into `model` module
