+++ backend/src/api/model/series.rs
@@ -18,6 +18,7 @@ use crate::{
     },
     db::util::{impl_from_db, select},
     model::{
+        AclForDb,
         AclItem,
         ExtraMetadata,
         Key,
@@ -46,7 +47,6 @@ use super::{
         define_sort_column_and_order,
         load_writable_for_user,
         BasicMetadata,
-        AclForDb,
         Connection,
         ConnectionQueryParts,
         PageInfo,
